* Add the missing clone function to C_CMMLDoc * Remove some warnings in my flac helper libraries * Use some size_t 's where appropriate. * Added a test to the CMML test app for the clone function.
git-svn-id: https://svn.xiph.org/trunk/oggdsf@8657 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ src/lib/codecs/flac/filters/dsfFLACDecoder/FLACDecodeOutputPin.cpp
@@ -61,7 +61,7 @@ HRESULT FLACDecodeOutputPin::CreateAndFillFormatBuffer(CMediaType* outMediaType,
 		locWaveFormat->wFormatTag = WAVE_FORMAT_PCM;
 		locWaveFormat->nChannels = locFilter->getFLACFormatBlock()->numChannels;
 		locWaveFormat->nSamplesPerSec =  locFilter->getFLACFormatBlock()->samplesPerSec;
-		locWaveFormat->wBitsPerSample = locFilter->getFLACFormatBlock()->numBitsPerSample;
+		locWaveFormat->wBitsPerSample = (WORD)locFilter->getFLACFormatBlock()->numBitsPerSample;
 		locWaveFormat->nBlockAlign = (locWaveFormat->nChannels) * (locWaveFormat->wBitsPerSample >> 3);
 		locWaveFormat->nAvgBytesPerSec = ((locWaveFormat->nChannels) * (locWaveFormat->wBitsPerSample >> 3)) * locWaveFormat->nSamplesPerSec;
 		locWaveFormat->cbSize = 0;
